Minimal refactor to StationHandler
