+++ Engine/Services/StationService.cs
@@ -12,16 +12,13 @@ using Core.Vehicles;
 using Core.Helper;
 
 /// <summary>
-/// Service responsible for managing the state of stations and chargers, handling events related to reservations, arrivals, and charging sessions.
+/// Coordinates station handlers and manages cross-station state (reservations, charger routing).
 /// </summary>
 public class StationService : IStationService
 {
-    private readonly Dictionary<int, (ChargerBase State, IChargerHandler Handler)> _chargerIndex = [];
-    private readonly Dictionary<ushort, Station> _stationIndex = [];
-    private readonly Dictionary<int, uint> _arrivalTimes = [];
+    private readonly Dictionary<ushort, StationHandler> _stationHandlers = [];
+    private readonly Dictionary<int, ushort> _chargerToStation = [];
     private readonly Dictionary<int, ushort> _evReservations = [];
-    private readonly EventScheduler _scheduler;
-    private readonly EVStore _eVStore;
 
     /// <summary>
     /// Initializes a new instance of the <see cref="StationService"/> class.
@@ -38,35 +35,25 @@ public class StationService : IStationService
         EVStore evStore,
         MetricsService metrics)
     {
-        _scheduler = scheduler;
-        _eVStore = evStore;
-
         foreach (var station in stations)
         {
-            _stationIndex[station.Id] = station;
+            var handler = new StationHandler(station, integrator, scheduler, evStore, metrics);
+            _stationHandlers[station.Id] = handler;
 
-            foreach (var charger in station.Chargers)
-            {
-                IChargerHandler handler = charger switch
-                {
-                    SingleCharger s => new SingleChargerHandler(s, integrator, scheduler, metrics),
-                    DualCharger d => new DualChargerHandler(d, integrator, scheduler, metrics),
-                    _ => throw Log.Error(0, 0, new InvalidOperationException($"Unknown charger type: {charger.GetType()}"), ((string Key, object Value))("Charger", charger))
-                };
-                _chargerIndex[charger.Id] = (charger, handler);
-            }
+            foreach (var chargerId in handler.ChargerIds)
+                _chargerToStation[chargerId] = station.Id;
         }
     }
 
     /// <inheritdoc/>
     public Station GetStation(ushort stationId)
-        => _stationIndex.TryGetValue(stationId, out var station)
-            ? station
+        => _stationHandlers.TryGetValue(stationId, out var handler)
+            ? handler.Station
             : throw Log.Error(0, 0, new SkillissueException($"Trying to get station {stationId} which does not exist."), ((string Key, object Value))("StationId", stationId));
 
     public IChargerHandler GetChargerHandler(int chargerId)
-        => _chargerIndex.TryGetValue(chargerId, out var pair)
-            ? pair.Handler
+        => _chargerToStation.TryGetValue(chargerId, out var stationId)
+            ? _stationHandlers[stationId].GetChargerHandler(chargerId)
             : throw new SkillissueException($"Trying to get charger handler {chargerId} which does not exist.");
 
     /// <summary>Gets the stationId that an EV has a reservation for if any.</summary>
@@ -101,37 +88,7 @@ public class StationService : IStationService
     /// </summary>
     /// <param name="e">The arrival event.</param>
     public void HandleArrivalAtStation(ArriveAtStation e)
-    {
-        ref var evRef = ref _eVStore.Get(e.EVId);
-        evRef.Advance(e.Time);
-        Log.Verbose(e.EVId, e.Time, $"Handling ArrivalAtStation for EV {e.EVId} at time {e.Time}. Current EV data: {evRef}. SoC: {evRef.Battery.StateOfCharge}. Wants to charge to {e.TargetSoC}");
-        var chargers = GetStation(e.StationId).Chargers;
-
-        if (evRef.Battery.StateOfCharge >= e.TargetSoC)
-            throw Log.Error(e.EVId, e.Time, new SkillissueException($"EV wants to charge to a SoC: {e.TargetSoC}, which is lower than its current SoC: {evRef.Battery.StateOfCharge}."), ((string Key, object Value))("EV", evRef), ((string Key, object Value))("TargetSoC", e.TargetSoC));
-
-        var target = chargers
-            .OrderBy(cs => cs.IsFree ? 0 : 1)
-            .ThenBy(cs => cs.Queue.Count)
-            .FirstOrDefault()
-            ?? throw Log.Error(e.EVId, e.Time, new SkillissueException($"Logic Error: Station {e.StationId} has no chargers."), ((string Key, object Value))("StationId", e.StationId));
-
-        var connectedEV = new ConnectedEV(
-            EVId: e.EVId,
-            CurrentSoC: evRef.Battery.StateOfCharge,
-            TargetSoC: e.TargetSoC,
-            CapacityKWh: evRef.Battery.MaxCapacityKWh,
-            MaxChargeRateKW: evRef.Battery.MaxChargeRateKW,
-            ArrivalTime: e.Time);
-
-        _arrivalTimes[e.EVId] = e.Time;
-        target.Queue.Enqueue(connectedEV);
-        _eVStore.Get(e.EVId).EVState = EVState.Queueing;
-        target.UpdateWindowStats();
-
-        if (target.IsFree)
-            StartChargingNextCar(target, e.Time, e.StationId);
-    }
+        => GetStationHandler(e.StationId).HandleArrivalAtStation(e);
 
     /// <summary>
     /// Called when a charging session ends for a specific EV.
@@ -140,103 +97,25 @@ public class StationService : IStationService
     /// <param name="e">The EndCharging event containing the EVId, ChargerId, and Time of the event.</param>
     public void HandleEndCharging(EndCharging e)
     {
-        if (!_chargerIndex.TryGetValue(e.ChargerId, out var entry))
+        if (!_chargerToStation.TryGetValue(e.ChargerId, out var stationId))
             return;
 
-        var (charger, handler) = entry;
-        charger.AccumulateEnergy(e.Time);
+        _stationHandlers[stationId].HandleEndCharging(e);
 
-        var finalSoC = handler.EndSession(e.EVId, e.Time);
-
-        ref var ev = ref _eVStore.Get(e.EVId);
-
-        if (finalSoC is { } soc)
-            ev.Battery.StateOfCharge = (float)Math.Clamp(soc, 0d, 1d);
-
-        if (!_arrivalTimes.TryGetValue(e.EVId, out var arrivalTime))
-            throw Log.Error(e.EVId, e.Time, new SkillissueException($"Logic Error: Missing arrival time for EV {e.EVId} at EndCharging."));
-
-        _arrivalTimes.Remove(e.EVId);
-        var timeAtStation = e.Time - arrivalTime;
-        ev.EVState = EVState.Driving;
-
-        if (ev.CanCompleteJourney(timeAtStation, ev.Preferences.MinAcceptableCharge))
-        {
-            Log.Info(e.EVId, e.Time, $"EV {e.EVId} has completed its charging and can continue to its destination with SoC {ev.Battery.StateOfCharge}. Scheduling arrival at destination.");
-            _scheduler.ScheduleEvent(new ArriveAtDestination(e.EVId, e.Time));
-        }
-        else
-        {
-            Log.Info(e.EVId, e.Time, $"EV {e.EVId} has completed its charging but cannot continue to its destination with SoC {ev.Battery.StateOfCharge}. Scheduling search for candidate stations.");
-            _scheduler.ScheduleEvent(new FindCandidateStations(e.EVId, ev.TimeAtNextFindCandidateCheck(e.Time)));
-        }
-
-        if (!_evReservations.TryGetValue(e.EVId, out var stationId))
+        if (!_evReservations.ContainsKey(e.EVId))
             throw Log.Error(e.EVId, e.Time, new SkillissueException("Should have a reservation at this point"));
 
         CancelReservation(e.EVId);
-        StartChargingNextCar(charger, e.Time, stationId);
-    }
-
-    private void StartChargingNextCar(ChargerBase charger, Time simNow, ushort stationId)
-    {
-        if (!charger.Queue.TryPeek(out var top))
-        {
-            return;
-        }
-
-        Log.Verbose(top.EVId, simNow, $"Starting next charge on charger {charger.Id} at station {stationId} at time {simNow}. Next EV in queue: {top.EVId}");
-        charger.AccumulateEnergy(simNow);
-        _chargerIndex[charger.Id].Handler.StartNext(simNow, stationId);
-        _eVStore.Get(top.EVId).EVState = EVState.Charging;
-        charger.UpdateWindowStats();
     }
 
     /// <inheritdoc/>
     public Time ExpectedWaitTime(ushort stationId, Time simNow, Time arrival)
-    {
-        var station = GetStation(stationId);
-        var waitTimes = new PriorityQueue<int, Time>();
-        var reservationQueues = new Dictionary<int, List<ConnectedEV>>();
-        var initialAvailability = new Dictionary<int, Time>();
+        => GetStationHandler(stationId).ExpectedWaitTime(simNow, arrival);
 
-        foreach (var charger in station.Chargers)
-        {
-            var handler = _chargerIndex[charger.Id].Handler;
-            var (availableAt, _) = handler.EstimateWaitTime(simNow);
-            var estimatedWait = availableAt + simNow;
-
-            waitTimes.Enqueue(charger.Id, estimatedWait);
-            reservationQueues[charger.Id] = [];
-            initialAvailability[charger.Id] = estimatedWait;
-        }
-
-        var reservationsBefore = station.Reservations.ReservationsBefore(arrival);
-
-        foreach (var reservation in reservationsBefore)
-        {
-            waitTimes.TryDequeue(out var chargerId, out var currentAvailableAt);
-
-            var battery = _eVStore.Get(reservation.EVId).Battery;
-
-            var connectedEV = new ConnectedEV(
-                EVId: reservation.EVId,
-                CurrentSoC: reservation.SoCAtArrival,
-                TargetSoC: reservation.TargetSoC,
-                CapacityKWh: battery.MaxCapacityKWh,
-                MaxChargeRateKW: battery.MaxChargeRateKW,
-                ArrivalTime: currentAvailableAt);
-
-            var queue = reservationQueues[chargerId];
-            queue.Add(connectedEV);
-
-            var handler = _chargerIndex[chargerId].Handler;
-            var baseTime = initialAvailability[chargerId];
-
-            var (availableAt, _) = handler.EstimateWaitTime(baseTime, queue);
-            waitTimes.Enqueue(chargerId, availableAt + baseTime);
-        }
-
-        return waitTimes.TryDequeue(out var _, out var minWait) ? minWait : simNow;
-    }
+    private StationHandler GetStationHandler(ushort stationId)
+        => _stationHandlers.TryGetValue(stationId, out var handler)
+            ? handler
+            : throw Log.Error(0, 0,
+                new SkillissueException($"Trying to get station handler {stationId} which does not exist."),
+                ((string Key, object Value))("StationId", stationId));
 }
